+++ nativelink-store/src/shard_store.rs
@@ -84,7 +84,7 @@ impl ShardStore {
         }))
     }
 
-    fn get_store_index(&self, store_key: &StoreKey) -> usize {
+    fn get_store_index(&self, store_key: &StoreKey) -> u64 {
         let key = match store_key {
             StoreKey::Digest(digest) => {
                 // Quote from std primitive array documentation:
@@ -155,7 +155,7 @@ impl StoreDriver for ShardStore {
                 .await
                 .err_tip(|| "In ShardStore::has_with_results() for store {store_idx}}");
         }
-        type KeyIdxVec = Vec<usize>;
+        type KeyIdxVec = Vec<u64>;
         type KeyVec<'a> = Vec<StoreKey<'a>>;
         let mut keys_for_store: Vec<(KeyIdxVec, KeyVec)> = self
             .weights_and_stores
